Restrict synchronizer usage on field count and stencil format
Different stages that use an identical communication pattern (number of
fields and stencil) use the same synchronizer at the moment.  This is
not problematic for synchronous execution of stages.  For asynchronous
stage execution, an additional stage identifier is required for a unique
synchronizer per stage.  Communicated data will be corrupted otherwise.