advection: add vfccl_stat to containt GetColor() from vof solver
XXX Field `vfccl_stat` may have value kClNone even in cells with u>0
close to the interface and should be used for statitics only,
and not for filtering the cells (e.g. in surface tension).
Required for single-field VOF solver where the criterion for merging the
cells is more strict to distinguish closely located bubbles.
Please enter the commit message for your changes. Lines starting